+++ Remnants/Patches/SpawnableScrapPatch.cs
@@ -27,12 +27,12 @@ namespace Remnants.Patches
         [HarmonyPrefix]
         public static void SpawnScrapInLevelPatch(RoundManager __instance)
         {
-            var mls = Remnants.Instance._mls;
+            var mls = Remnants.Instance.Mls;
             mls.LogInfo("In SpawnScrapInLevelPatch function!");
             //Here we will delete all items that are banned, currently replaced by itemsscrap for easy of testing
             List<SpawnableItemWithRarity> spawnableScrapList = __instance.currentLevel.spawnableScrap;
-            List<ScrapItemData> scrapItemDataList = ScrapDataListBehaviour.GetScrapItemDataList();
-            spawnableScrapList.RemoveAll(spawnableItem => scrapItemDataList.FindIndex(itemData => itemData.Isbanned && itemData.ScrapItemName == spawnableItem.spawnableItem.name) != -1);
+            List<RemnantData> scrapItemDataList = RemnantDataListBehaviour.GetRemnantDataList();
+            spawnableScrapList.RemoveAll(spawnableItem => scrapItemDataList.FindIndex(itemData => !itemData.ShouldSpawn && itemData.RemnantItemName == spawnableItem.spawnableItem.name) != -1);
             __instance.currentLevel.spawnableScrap = spawnableScrapList;
         }
         #endregion